+++ src/main/kotlin/store/domain/controller/InputController.kt
@@ -1,5 +1,6 @@
-package store.domain.input
+package store.domain.controller
 
+import store.domain.validator.InputValidator
 import store.model.OrderItem
 import store.model.Product
 import store.view.InputView
@@ -23,5 +24,13 @@ class InputController(private val inputView: InputView, private val inputValidat
 
     fun promptApplyMembershipDiscount(): Boolean {
         return inputView.promptApplyMembershipDiscount()
+    }
+
+    fun promptAddItemsForPromotion(productName: String, additionalQuantityNeeded: Int): Boolean {
+        return inputView.promptAddItemsForPromotion(productName, additionalQuantityNeeded)
+    }
+
+    fun promptPayFullPriceForShortage(): Boolean {
+        return inputView.promptPayFullPriceForShortage()
     }
 }
